% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{as.gsDesign}
\alias{as.gsDesign}
\title{Convert fit to a \code{gsDesign} object}
\usage{
as.gsDesign(x)
}
\arguments{
\item{x}{an object of class \code{gsDesignOC}}
}
\value{
an object of class \code{gsDesign}
}
\description{
The \code{as.gsDesign} converts a \code{gsDesignOC} object to a form that works with the
\code{gsDesign} package
}
\examples{

require(gsDesign)
g <- gsDesignOC(n.stages=2, rE.seq=c(2,1))
gg <- as.gsDesign(g)
gg
}
\author{
Aniko Szabo
}
\keyword{design}
\keyword{nonparametric}
