% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{calc.bounds}
\alias{calc.bounds}
\title{Calculate the information times and efficacy/futility boundary values given the alpha-spending sequence and desired operating characteristics}
\usage{
calc.bounds(x, alpha.seq)
}
\arguments{
\item{x}{an object of class \code{gsDesignOC}}
}
\description{
Calculate the information times and efficacy/futility boundary values given the alpha-spending sequence and desired operating characteristics
}
