% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{convert.bounds}
\alias{convert.bounds}
\alias{convert.bounds2}
\title{Conversion between nominal significance levels and alternative hypothesis values}
\usage{
convert.bounds(nominal.level = NULL, theta = NULL, n, power, theta.null)

convert.bounds2(
  nominal.level = NULL,
  theta = NULL,
  n,
  power,
  theta.null,
  power.futility,
  theta.alt
)
}
\arguments{
\item{nominal.level}{numeric vector (for \code{convert.bounds}) or
numeric matrix with two rows (for \code{convert.bounds2}) of the alternative hypotheses for each stage.
When a matrix, row 1 corresponds to the efficacy boundary and row 2 to the futility. Defaults to NULL.}

\item{n}{integer vector of sample sizes of each stage. Its sum is the total study sample size.}

\item{power}{numeric, the target power at each stage.}

\item{theta.null}{numeric, the null hypothesis being tested}

\item{power.futility}{numeric, the target power for futility testing at each stage.}

\item{theta.alt}{numeric, the alternative hypothesis being tested for the futility boundary.}
}
\value{
a list with all the inputs, with the NULL component filled in
}
\description{
The \code{convert.bounds} function performs conversion between nominal significance levels
for testing eqn{H_0: \theta=\theta_0}{H0: theta=theta0} and the values for the alternatives that have a prespecified power
when \eqn{H_0}{H0} is tested at the nominal level.
}
\details{
The \code{convert.bounds2} extends this function for both and efficacy and futility boundary,
where for the futility boundary the hypothesis eqn{H_A: \theta=\theta_A}{HA: theta=thetaA}  is tested.
The power is cumulative over the previous stages.
Either the first 2 parameters has to be specified, and the other one will be computed.
}
\section{Functions}{
\itemize{
\item \code{convert.bounds2}: Conversion between nominal significance and alternatives with both futility and efficacy
}}

\examples{

(c1 <- convert.bounds(nominal.level = c(0.01, 0.02),
                      n = c(30, 50), power=0.8, theta.null=0))
convert.bounds(theta = c1$theta, n = c(30, 50), power=0.8, theta.null=0)$nominal.level
(c2 <- convert.bounds2(nominal.level = rbind(c(0.01, 0.02), c(0.05,0.1)),
                      n = c(30, 50), power=0.8, theta.null=0, power.futility=0.9, theta.alt=0.2))
convert.bounds2(theta = c2$theta, n = c(30, 50), power=0.8, theta.null=0,
               power.futility=0.9, theta.alt=0.2)$nominal.level
}
\author{
Aniko Szabo
}
\keyword{internal}
