% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{gsDesignOC}
\alias{gsDesignOC}
\title{Find optimal group-sequential design}
\usage{
gsDesignOC(
  thA,
  thA.seq,
  th0 = 0,
  th0.seq = NULL,
  sig.level = 0.025,
  power = 0.9,
  power.efficacy = power,
  power.futility = power,
  futility.type = c("none", "non-binding", "binding"),
  mix.theta = thA,
  mix.w = rep(1, length(mix.theta)),
  control = ocControl()
)
}
\arguments{
\item{thA}{numeric; effect size under the alternative hypothesis}

\item{thA.seq}{monotone numeric vector of values getting closer to thA from outside the
th0-thA range (ie, if thA > th0, they should form a decreasing sequence with all values > thA). #'For the k-th value the study will stop for efficacy at or before the k-th stage with probability \code{power.efficacy}.}

\item{th0}{numeric; effect size under the null hypothesis}

\item{th0.seq}{monotone numeric vector of values getting closer to th0 from outside the
th0-thA range (ie, if thA > th0, they should form an increasing sequence with all values < th0).
The study will stop for futility at or before the k-th stage with probability \code{power.futility}.
Its length should be equal to that of \code{thA.seq}. The default value of \code{NULL} implies no
futility monitoring.}

\item{sig.level}{numeric; the one-sided significance level. The default is 0.025.}

\item{power}{numeric; the desired study power. The default is 0.9.}

\item{power.efficacy}{numeric; the probability of stopping for efficacy at stage k under \code{thA.seq}}

\item{power.futility}{numeric; the probability of stopping for futility at stage k under \code{th0.seq}}

\item{futility.type}{character string, one of \code{c("none", "non-binding","binding")} or their
unique abbreviations. Specifies whether a futility boundary should not be used at all ("none"), or if it
is used, then whether the effect of the futility boundary should be included
in the efficacy power/type I error calculations ("binding"), or not ("non-binding").}

\item{mix.theta}{numeric vector; specifies the set of alternatives under which the design is optimized.
Defaults to \code{thA}.}

\item{mix.w}{numeric vector of length equal to that of \code{mix.theta}. The weights of the
elements of \code{mix.theta} in the optimality criterion. It will be normalized to a sum of 1.
Defaults to equal weights.}

\item{control}{list of parameters controlling the estimation altorithm to replace the default
values returned by the \code{glsControl} function.}
}
\value{
an object of class \code{gsDesignOC}
}
\description{
The \code{gsDesignOC} function finds the analysis times, boundaries, and sample size
for a group-sequential trial
}
\examples{

gsDesignOC(0.3, thA.seq = c(1, 0.5))

}
\references{
Szabo, A, Tarima, S (?) Operating-characteristic guided design of group-sequential trials.
}
\author{
Aniko Szabo
}
\keyword{design}
\keyword{nonparametric}
