% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nom.to.bnd}
\alias{nom.to.bnd}
\alias{bnd.to.nom}
\title{Conversion between boundary and nominal significance level}
\usage{
nom.to.bnd(nominal.level, n, theta.null, theta.alt = NULL)

bnd.to.nom(cutoffs, n, theta.null, theta.alt = NULL)
}
\arguments{
\item{nominal.level}{numeric vector or matrix with two rows of the nominal significance level for each stage.
When a matrix, row 1 corresponds to the efficacy boundary and row 2 to the futility. Defaults to NULL.}

\item{n}{integer vector of sample sizes of each stage. Its sum is the total study sample size.}

\item{theta.null}{numeric, the null hypothesis being tested}

\item{theta.alt}{numeric, the alternative hypothesis being tested for the futility bound calculation}

\item{cutoffs}{numeric vector or numeric matrix with two rows of the z-test cutoffs for each stage.
When a matrix, row 1 corresponds to the efficacy boundary and row 2 to the futility. Defaults to NULL.}
}
\description{
The \code{nom.to.bnd} and \code{bnd.to.nom} functions perform conversion between the boundary and
matching significance level.
}
\section{Functions}{
\itemize{
\item \code{bnd.to.nom}: Conversion from nominal significance level to boundary
}}

\examples{

## only efficacy
(b1 <- nom.to.bnd(nominal.level = c(0.01, 0.02), n = c(30, 50)))
bnd.to.nom(b1, n=c(30,50))
## both efficacy and futility
(b2 <- nom.to.bnd(nominal.level = rbind(c(0.01, 0.02), c(0.05,0.1)),
                n = c(30, 50)))
bnd.to.nom(b2, n=c(30,50))
}
\keyword{internal}
