% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{oc}
\alias{oc}
\title{Operating characteristics of a potential design}
\usage{
oc(x)
}
\arguments{
\item{x}{object of class \code{gsDesignOC}}
}
\value{

}
\description{
The values supplied in the function call replace the defaults and a list with all possible #'arguments is returned. The returned list is used as the \code{control} argument to the
\code{gsDesignOC} function.
}
\examples{


}
\author{
Aniko Szabo
}
\keyword{design}
