% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{ocControl}
\alias{ocControl}
\title{Control values for gsDesignOC}
\usage{
ocControl(optim.method = c("direct", "dynamic"))
}
\arguments{
\item{optim.method}{character; defines the optimization method used: \code{"dynamic"} uses a recursive dynamic algorithm #' which selects the alpha-spending one stage at a time, \code{"direct"} uses the "Nelder-Mead" algorithm from
\code{\link{optim}} to simultaneously search over all possible alpha-spending sequences.}
}
\value{
a list with components for each of the possible arguments
}
\description{
The values supplied in the function call replace the defaults and a list with all possible
arguments is returned. The returned list is used as the \code{control} argument to the
\code{gsDesignOC} function.
}
\examples{

ocControl(optim.method = "direct")
}
\author{
Aniko Szabo
}
\keyword{design}
