% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{overall.crossprob}
\alias{overall.crossprob}
\title{Overall probability of crossing the efficacy and futility boundaries}
\usage{
overall.crossprob(
  theta,
  n.I,
  n,
  thA.seq,
  power,
  th0.seq = NULL,
  power.futility = NULL,
  sig.level_final = NULL
)
}
\arguments{
\item{theta}{numeric; value of the parameter under which the crossing probabilities are computed}

\item{n.I}{numeric; cumulative proportio}

\item{thA.seq}{monotone numeric vector of values getting closer to thA from outside the
th0-thA range (ie, if thA > th0, they should form a decreasing sequence with all values > thA). #'For the k-th value the study will stop for efficacy at or before the k-th stage with probability \code{power}.}

\item{power}{numeric; the desired study power. The default is 0.9. This value will also be
used to set the probability of stopping for efficacy at stage k under \code{thA.seq}.}

\item{th0.seq}{monotone numeric vector of values getting closer to th0 from outside the
th0-thA range (ie, if thA > th0, they should form an increasing sequence with all values < th0).
The study will stop for futility at or before the k-th stage with probability \code{power.futility}.
Its length should be equal to that of \code{thA.seq}. The default value of \code{NULL} implies no
futility monitoring.}

\item{power.futility}{numeric; the probability of stopping for futility at stage k under \code{th0.seq}}

\item{sig.level_final}{numeric; the desired nominal significance level for testing the null
hypothesis at the final stage. Should be between 0 and \code{sig.level}. If NULL, the value
will be found using the optimization criterion.}
}
\value{
a list with all the inputs, with the NULL component filled in
}
\description{
The \code{overall.crossprob} calculates the probability of crossing the efficacy and, optionally,
the futility bound, given a series of alternative (and optionally, null) hypotheses tested at
each stage.
}
\examples{

(c1 <- convert.bounds(nominal.level = c(0.01, 0.02),
                      n = c(30, 50), power=0.8, theta.null=0))
convert.bounds(theta = c1$theta, n = c(30, 50), power=0.8, theta.null=0)$nominal.level
}
\author{
Aniko Szabo
}
\keyword{internal}
