% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{ztest.I}
\alias{ztest.I}
\title{Fixed sample-size information for one-sided test}
\usage{
ztest.I(delta, sig.level, power)
}
\arguments{
\item{delta}{numeric; targeted standardized effect size}

\item{sig.level}{numeric; one-sided significance level}

\item{power}{numeric; target power}
}
\value{
numeric value with the (fractional) information achieving the target power
}
\description{
The \code{ztest.I} function finds the information required of a one-sided Z-test with given power
and significance level
}
\examples{

ztest.I(delta=1, sig.level=0.05, power=0.9)
}
\author{
Aniko Szabo
}
\keyword{internal}
