% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{ztest.n}
\alias{ztest.n}
\title{One-sample Z-based sample size for one-sided test}
\usage{
ztest.n(delta, sd, sig.level, power)
}
\arguments{
\item{delta}{numeric; targeted effect size}

\item{sd}{numeric; standard deviation, assumed known}

\item{sig.level}{numeric; one-sided significance level}

\item{power}{numeric; target power}
}
\value{
numeric value with the (fractional) sample size achieving the target power
}
\description{
The \code{ztest.n} function finds the sample size of a one-sided Z-test with given power
and significance level
}
\examples{

ztest.n(delta=1, sd=1, sig.level=0.05, power=0.9)
}
\author{
Aniko Szabo
}
\keyword{internal}
