% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{calc.bounds}
\alias{calc.bounds}
\title{Calculate the information times and efficacy/futility boundary values given the alpha-spending sequence and desired operating characteristics}
\usage{
calc.bounds(x, alpha.seq)
}
\arguments{
\item{x}{a list with desired operating characteristics. Should have \code{rE.seq}, \code{sig.level},
\code{power}, \code{power.efficacy}, and \code{futility.type} components; and \code{rF.seq} and \code{power.futility}
components if lower boundary is requested. Objects of class \code{gsDesignOC} (potentially incomplete)  have
these components.}

\item{alpha.seq}{numeric vector of stage-specific alpha-spending; values should add up to \code{x$sig.level}.}
}
\value{
The value of \code{x} augmented with the following components:
\describe{
\item{info}{numeric vector; information times for the analyses}
\item{lower}{numeric vector; lower Z-score boundary for the futility decisions}
\item{upper}{numeric vector; upper Z-score boundary for the efficacy decisions}
\item{spending}{numeric vector}{the value of alpha-spending sequence \code{alpha.seq}}
}
}
\description{
Calculate the information times and efficacy/futility boundary values given the alpha-spending sequence and desired operating characteristics
}
\keyword{internal}
