% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{gsDesignOC}
\alias{gsDesignOC}
\alias{print.gsDesignOC}
\title{Find optimal group-sequential design}
\usage{
gsDesignOC(
  n.stages,
  rE.seq,
  rF.seq = NULL,
  n.fix = 1,
  sig.level = 0.025,
  power = 0.9,
  power.efficacy = power,
  power.futility = power,
  futility.type = c("none", "non-binding", "binding"),
  r_EN = 1,
  r_EN.w = rep(1, length(r_EN)),
  spending = NULL,
  control = ocControl()
)

\method{print}{gsDesignOC}(x, ...)
}
\arguments{
\item{n.stages}{integer; number of stages planned. One stage implies no interim analysis.}

\item{rE.seq}{monotone decreasing numeric vector of length \code{n.stages} ending with 1.
If it has length \code{n.stages-1}, a 1 will be appended to the end.
For the k-th interim value the study will stop for efficacy at or before the k-th stage with probability
\code{power.efficacy} if the true effect is \code{rE.seq[k]} times the effect under the alternative hypothesis.}

\item{rF.seq}{monotone increasing numeric vector of length \code{n.stages} ending with 0.
If it has length \code{n.stages-1}, a 0 will be appended to the end.
The study will stop for futility at or before the k-th stage with probability \code{power.futility}.
if the true effect is \code{rE.seq[k]} times the effect under the alternative hypothesis, ie
the true effect is actually in the opposite direction of the hypothesized effect.
The default value of \code{NULL} implies no futility monitoring.}

\item{n.fix}{numeric; the sample size for a one-stage design without interim tests. Defaults to 1,
in which case the resulting sample sizes can be interpreted as being relative to the single-stage study sample size.}

\item{sig.level}{numeric; the one-sided significance level. The default is 0.025.}

\item{power}{numeric; the desired study power. The default is 0.9.}

\item{power.efficacy}{numeric; the probability of stopping for efficacy at stage k under \code{rE.seq}}

\item{power.futility}{numeric; the probability of stopping for futility at stage k under \code{rF.seq}}

\item{futility.type}{character string, one of \code{c("none", "non-binding","binding")} or their
unique abbreviations. Specifies whether a futility boundary should not be used at all ("none"), or if it
is used, then whether the effect of the futility boundary should be included
in the efficacy power/type I error calculations ("binding"), or not ("non-binding").}

\item{r_EN}{numeric vector; specifies the set of alternatives under which the design is optimized.
It is interpreted multiplicatively compared to the design alternative hypothesis.
Defaults to 1, implying minimization under the alternative hypothesis.}

\item{r_EN.w}{numeric vector of length equal to that of \code{r_EN}. The weights of the
elements of \code{r_EN} in the optimality criterion. It will be normalized to a sum of 1.
Defaults to equal weights.}

\item{spending}{(optional) numeric vector of stage-wise alpha-spending with positive values
summing to \code{sig.level} used as starting values for direct optimization. If used in conjunction with
\code{optim.method="none"} within \code{ocControl}, a non-optimized design using this alpha-spending
will be returned. The starting value will be selected automatically for the default value of NULL.}

\item{control}{list of parameters controlling the estimation altorithm to replace the default
values returned by the \code{glsControl} function.}
}
\value{
an object of class \code{gsDesignOC}
}
\description{
The \code{gsDesignOC} function finds the analysis times, boundaries, and sample size
for a group-sequential trial
}
\examples{

gsDesignOC(2, rE.seq = c(1.5,1), rF.seq = c(-0.5,0),
          power.efficacy=0.8, power.futility=0.8, power=0.9,
          futility.type = "non-binding")

## pre-specifiy alpha-spending
gsDesignOC(2, rE.seq = c(1.5,1), rF.seq = c(-0.5,0),
          power.efficacy=0.8, power.futility=0.8, power=0.9,
          futility.type = "non-binding", spending=c(0.005, 0.02),
          control = ocControl(optim.method="none"))

}
\references{
Szabo, A, Tarima, S (?) Operating-characteristic guided design of group-sequential trials.
}
\author{
Aniko Szabo
}
\keyword{design}
\keyword{nonparametric}
