% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{oc}
\alias{oc}
\title{Operating characteristics of a potential design}
\usage{
oc(x, EN_theta = x$thetaA, mix.w = rep(1, length(EN_theta)))
}
\arguments{
\item{x}{object of class \code{gsDesignOC}}

\item{EN_theta}{numeric vector of parameter values for which expected sample size calculation is performed}

\item{mix.w}{numeric vector of positive weights for each value of \code{EN_theta}. Defaults to equal weights.}
}
\value{
a list with the following elements:
\describe{
\item{ave.EN}{numeric; weighted average of expected sample sizes under the requested alternatives}
\item{EN.vec}{numeric vector; expected sample sizes under each of the requested alternatives}
}
}
\description{
Calculates the average expected information under a weighted set of values for the alternative hypothesis.
}
\examples{


}
\author{
Aniko Szabo
}
\keyword{design}
