% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{oc}
\alias{oc}
\title{Operating characteristics of a potential design}
\usage{
oc(x, EN_theta = x$thA, mix.w = rep(1, length(EN_theta)))
}
\arguments{
\item{x}{object of class \code{gsDesignOC}}

\item{EN_theta}{numeric vector of parameter values for which expected sample size calculation is performed}

\item{mix.w}{numeric vector of positive weights for each value of \code{EN_theta}. Defaults to equal weights.}
}
\value{
a list with the following elements:
\describe{
\item{ave.EN}{numeric; weighted average of expected sample sizes under the requested alternatives}
\item{EN.vec}{numeric vector; expected sample sizes under each of the requested alternatives}
\item{thA.cumcross}{numeric vector; cumulative probability of stopping for efficacy at or before stage k
 under thA.seq[k], including thA at the end.}
\item{th0.cumcross}{numeric vector}{numeric vector; cumulative probability of stopping for futility at or before stage k
 under th0.seq[k], including th0 at the end.}
}
}
\description{
Calculates the average expected information under a weighted set of values for the alternative hypothesis, and
the probability of stopping for futility or effecacy under the design alternatives. If the futility boundary
is non-binding, then the lower boundary is not included in the calculation of the expected sample size and
efficacy stopping probabilities, but it is included in the futility stopping calculations.
}
\examples{
g <- gsDesignOC(0.3, thA.seq = 0.5, th0.seq = -0.3, power.efficacy=0.8,
          power.futility=0.8, power=0.9,
          futility.type = "non-binding", mix.theta=c(0.5, 0.3, 0))
oc(g, EN_theta = c(0.5, 0.3, 0))


}
\author{
Aniko Szabo
}
\keyword{design}
