% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignOC.R
\name{ocControl}
\alias{ocControl}
\title{Control values for gsDesignOC}
\usage{
ocControl(optim.penalty = 1000)
}
\arguments{
\item{optim.penalty}{numeric; relative weight of terms ensuring target type I and type II
errors versus sample size in optimization routine}
}
\value{
a list with components for each of the possible arguments
}
\description{
The values supplied in the function call replace the defaults and a list with all possible
arguments is returned. The returned list is used as the \code{control} argument to the
\code{gsDesignOC} function.
}
\examples{

ocControl(optim.penalty = 100)
}
\author{
Aniko Szabo
}
\keyword{design}
